Calculate Projection subheaders
We’re now calculating the values in the ‘Projection’ section’s ‘average
rate’ and ‘years until’ subheaders (using the real data returned by the
api).

Graph aside (it currently just displays the exact same stuff as the
‘all’ section), note that we’ll likely need to change the hardcoded
tipping point of ‘530ppm’ as it’s just one of the various important
‘threshold’ points listed in the SPM.1 table.

(530ppm is notable for being the point where we’re <= 50% chance of
staying below a 2C increase, but there is like a different point that’d
be more suitable. If there’s a discrete point that starts runaway
release of permafrost-trapped methane, that might be a good one?)

Note that this still needs unit tests for the new
functionality/calculations.